Rewrite in Go
This does roughly double the amount of code (if you ignore the massive
generated go.mod and go.sum files), but it produces a self-contained
binary I can run without needing to either configure a Python
environment or stick everything in a Docker container.

And also the code is very straightforward - the largeness is just
because Go is verbose.